fix: resolve static-analysis findings (zero warnings/errors)
Production fixes:
- GiPass: range-for over the ping-pong reservoir Buffer[2] arrays.
- Application: mark destructor override; use std::trunc instead of int-cast
  rounding in the window-title FPS/EV formatting (avoids int->float narrowing).
- GpuCullPass / IblPrecompute / LightCuller: value-initialise the push-constant
  struct members (cppcheck uninitMemberVarNoCtor) — layout unchanged, CPU
  overwrites them before vkCmdPushConstants.
- IblPrecompute: fix a malformed cppcheck inline-suppression comment (a
  non-ASCII em-dash was parsed as a bad suppression attribute).

Test:
- Widen before multiplying in the env-sampling CDF reserve.

clang-tidy + cppcheck now emit zero warnings and zero errors; 17 tests green.

+++ tests/module/test_env_sampling_contract.cpp
@@ -126,7 +126,7 @@ TEST(TheiaEnvSamplingContract, PdfEvaluationMatchesSampledDirections) {
     constexpr uint32_t H = 4U;
     const std::vector<float> marginal{0.0F, 0.1F, 0.35F, 0.7F, 1.0F};
     std::vector<float> conditional;
-    conditional.reserve(H * (W + 1U));
+    conditional.reserve(static_cast<size_t>(H) * (W + 1U));
     for (uint32_t row = 0U; row < H; ++row) {
         conditional.push_back(0.0F);
         for (uint32_t col = 0U; col < W; ++col) {
